+++ bill/calculator.go
@@ -13,58 +13,27 @@ import (
 	"github.com/invopop/validation"
 )
 
-// billable defines the methods required to be able to perform calculations and
-// other operations on a bill document with a common basic structure.
-type billable interface {
-	// From tax.Regime
-	RegimeDef() *tax.RegimeDef
-
-	// From tax.Tags
-	HasTags(tags ...cbc.Key) bool
-	GetTags() []cbc.Key
-
-	getIssueDate() cal.Date
-	getIssueTime() *cal.Time
-	getValueDate() *cal.Date
-	getTax() *Tax
-	getPreceding() []*org.DocumentRef
-	getCustomer() *org.Party
-	getCurrency() currency.Code
-	getExchangeRates() []*currency.ExchangeRate
-	getLines() []*Line
-	getDiscounts() []*Discount
-	getCharges() []*Charge
-	getPaymentDetails() *PaymentDetails
-	getTotals() *Totals
-	getComplements() []*schema.Object
-
-	setIssueDate(cal.Date)
-	setIssueTime(*cal.Time)
-	setCurrency(currency.Code)
-	setTotals(*Totals)
-}
-
 func calculate(doc billable) error {
 	r := doc.RegimeDef() // may be nil!
 	calculateIssueDateAndTime(r, doc)
 
 	// Get the date used for tax calculations
-	date := doc.getValueDate()
+	date := doc.GetValueDate()
 	if date == nil {
-		id := doc.getIssueDate()
+		id := doc.GetIssueDate()
 		date = &id
 	}
 
 	// Convert empty or invalid currency to the regime's currency
-	if doc.getCurrency() == currency.CodeEmpty || doc.getCurrency().Def() == nil {
+	if doc.GetCurrency() == currency.CodeEmpty || doc.GetCurrency().Def() == nil {
 		if r == nil {
 			return validation.Errors{"currency": errors.New("missing")}
 		}
-		doc.setCurrency(r.Currency)
+		doc.SetCurrency(r.Currency)
 	}
-	cur := doc.getCurrency()
+	cur := doc.GetCurrency()
 
-	t := doc.getTotals()
+	t := doc.GetTotals()
 	// Prepare the totals we'll need with amounts based on currency
 	if t == nil {
 		t = new(Totals)
@@ -75,7 +44,7 @@ func calculate(doc billable) error {
 	// Figure out rounding rules and if prices include tax early
 	var pit cbc.Code
 	var rr cbc.Key
-	if tx := doc.getTax(); tx != nil {
+	if tx := doc.GetTax(); tx != nil {
 		if tx.PricesInclude != "" {
 			pit = tx.PricesInclude
 		}
@@ -93,59 +62,59 @@ func calculate(doc billable) error {
 	}
 
 	// Complements
-	if err := calculateComplements(doc.getComplements()); err != nil {
+	if err := calculateComplements(doc.GetComplements()); err != nil {
 		return validation.Errors{"complements": err}
 	}
 
 	// Preceding
-	calculateOrgDocumentRefs(doc.getPreceding(), cur, rr)
+	calculateOrgDocumentRefs(doc.GetPreceding(), cur, rr)
 
 	// Lines
-	if err := calculateLines(doc.getLines(), cur, doc.getExchangeRates(), rr); err != nil {
+	if err := calculateLines(doc.GetLines(), cur, doc.GetExchangeRates(), rr); err != nil {
 		return validation.Errors{"lines": err}
 	}
-	t.Sum = calculateLineSum(doc.getLines(), cur)
+	t.Sum = calculateLineSum(doc.GetLines(), cur)
 	t.Total = t.Sum
 
 	// Discount Lines
-	calculateDiscounts(doc.getDiscounts(), cur, t.Sum, rr)
-	if discounts := calculateDiscountSum(doc.getDiscounts(), cur); discounts != nil {
+	calculateDiscounts(doc.GetDiscounts(), cur, t.Sum, rr)
+	if discounts := calculateDiscountSum(doc.GetDiscounts(), cur); discounts != nil {
 		t.Discount = discounts
 		t.Total = t.Total.Subtract(*discounts)
 	}
 
 	// Charge Lines
-	calculateCharges(doc.getCharges(), cur, t.Sum, rr)
-	if charges := calculateChargeSum(doc.getCharges(), cur); charges != nil {
+	calculateCharges(doc.GetCharges(), cur, t.Sum, rr)
+	if charges := calculateChargeSum(doc.GetCharges(), cur); charges != nil {
 		t.Charge = charges
 		t.Total = t.Total.Add(*charges)
 	}
 
 	// Build list of taxable lines
 	tls := make([]tax.TaxableLine, 0)
-	for _, l := range doc.getLines() {
+	for _, l := range doc.GetLines() {
 		if l.Total != nil {
 			tls = append(tls, l)
 		}
 	}
-	for _, l := range doc.getDiscounts() {
+	for _, l := range doc.GetDiscounts() {
 		tls = append(tls, l)
 	}
-	for _, l := range doc.getCharges() {
+	for _, l := range doc.GetCharges() {
 		tls = append(tls, l)
 	}
 
 	if len(tls) == 0 {
 		// This applies for orders and deliveries that might not have
 		// any pricing details.
-		doc.setTotals(nil)
+		doc.SetTotals(nil)
 		return nil
 	}
 
 	// Now figure out the tax totals
 	t.Taxes = new(tax.Total)
 	tc := &tax.TotalCalculator{
-		Currency: doc.getCurrency(),
+		Currency: doc.GetCurrency(),
 		Rounding: rr,
 		Country:  r.GetCountry(),
 		Tags:     doc.GetTags(),
@@ -183,16 +152,16 @@ func calculate(doc billable) error {
 	// Before calculating the amount due and advances, we need to round
 	// everything. Payments reflect real monetary values and can never
 	// be fractions of the currency.
-	roundLines(doc.getLines())
-	roundDiscounts(doc.getDiscounts(), cur)
-	roundCharges(doc.getCharges(), cur)
+	roundLines(doc.GetLines())
+	roundDiscounts(doc.GetDiscounts(), cur)
+	roundCharges(doc.GetCharges(), cur)
 	t.round(zero)
 
 	if t.Rounding != nil {
 		// BT-144 in EN16931
 		t.Payable = t.Payable.Add(*t.Rounding)
 	}
-	if pd := doc.getPaymentDetails(); pd != nil {
+	if pd := doc.GetPaymentDetails(); pd != nil {
 		pd.calculateAdvances(zero, t.Payable)
 		// Deal with advances, if any. Note that in the current
 		// implementation multiple percentage advances are likely to
@@ -205,20 +174,20 @@ func calculate(doc billable) error {
 		// Calculate any due date amounts
 		pd.Terms.CalculateDues(zero, t.Payable)
 	}
-	doc.setTotals(t)
+	doc.SetTotals(t)
 
 	return nil
 }
 
 func calculateIssueDateAndTime(r *tax.RegimeDef, doc billable) {
 	tz := r.TimeLocation()
-	if doc.getIssueTime() != nil && doc.getIssueTime().IsZero() {
+	if doc.GetIssueTime() != nil && doc.GetIssueTime().IsZero() {
 		dn := cal.ThisSecondIn(tz)
 		tn := dn.Time()
-		doc.setIssueDate(dn.Date())
-		doc.setIssueTime(&tn)
-	} else if doc.getIssueDate().IsZero() {
-		doc.setIssueDate(cal.TodayIn(tz))
+		doc.SetIssueDate(dn.Date())
+		doc.SetIssueTime(&tn)
+	} else if doc.GetIssueDate().IsZero() {
+		doc.SetIssueDate(cal.TodayIn(tz))
 	}
 }
 
@@ -232,37 +201,37 @@ func calculateOrgDocumentRefs(drs []*org.DocumentRef, cur currency.Code, rr cbc.
 }
 
 func canRemoveIncludedTaxes(doc billable) bool {
-	return doc.getTax() != nil && !doc.getTax().PricesInclude.IsEmpty()
+	return doc.GetTax() != nil && !doc.GetTax().PricesInclude.IsEmpty()
 }
 
 func removeIncludedTaxes(doc billable) error {
 	if !canRemoveIncludedTaxes(doc) {
 		return nil
 	}
-	tpi := doc.getTax().PricesInclude
+	tpi := doc.GetTax().PricesInclude
 
-	totalWithTax := doc.getTotals().TotalWithTax
+	totalWithTax := doc.GetTotals().TotalWithTax
 
-	doc.setTotals(new(Totals))
-	lines := doc.getLines()
-	for i, l := range doc.getLines() {
+	doc.SetTotals(new(Totals))
+	lines := doc.GetLines()
+	for i, l := range doc.GetLines() {
 		lines[i] = removeLineIncludedTaxes(l, tpi)
 	}
 
-	discounts := doc.getDiscounts()
+	discounts := doc.GetDiscounts()
 	if len(discounts) > 0 {
 		for i, l := range discounts {
 			discounts[i] = l.removeIncludedTaxes(tpi)
 		}
 	}
-	charges := doc.getCharges()
+	charges := doc.GetCharges()
 	if len(charges) > 0 {
 		for i, l := range charges {
 			charges[i] = l.removeIncludedTaxes(tpi)
 		}
 	}
 
-	tx := doc.getTax()
+	tx := doc.GetTax()
 	tx.PricesInclude = ""
 
 	if err := calculate(doc); err != nil {
@@ -270,7 +239,7 @@ func removeIncludedTaxes(doc billable) error {
 	}
 
 	// Account for any rounding errors that we just can't handle
-	t := doc.getTotals()
+	t := doc.GetTotals()
 	if !totalWithTax.Equals(t.TotalWithTax) {
 		rnd := totalWithTax.Subtract(t.TotalWithTax)
 		t.Rounding = &rnd
@@ -283,17 +252,17 @@ func removeIncludedTaxes(doc billable) error {
 }
 
 func applyCustomerRates(doc billable) {
-	if doc.getCustomer() == nil || doc.getCustomer().TaxID == nil {
+	if doc.GetCustomer() == nil || doc.GetCustomer().TaxID == nil {
 		return
 	}
-	country := doc.getCustomer().TaxID.Country
-	for _, l := range doc.getLines() {
+	country := doc.GetCustomer().TaxID.Country
+	for _, l := range doc.GetLines() {
 		addCountryToTaxes(l.Taxes, country)
 	}
-	for _, d := range doc.getDiscounts() {
+	for _, d := range doc.GetDiscounts() {
 		addCountryToTaxes(d.Taxes, country)
 	}
-	for _, c := range doc.getCharges() {
+	for _, c := range doc.GetCharges() {
 		addCountryToTaxes(c.Taxes, country)
 	}
 }